agentic-rag/requirements.txt
crewai
crewai-tools
langchain
langchain-openai
langchain-community
python-dotenv
plotly
pandas
docling